Update SearchForm.isx (input in nav bar)
