+++ src/components/Header.jsx
@@ -2,13 +2,13 @@ import React from 'react'
 import HeaderStyled from './HeaderStyled'
 import Logo from './Logo'
 import Navigation from './Navigation'
-import SearchResult from './SearchResult'
+import SearchForm from './SearchForm'
 
 export default function Header() {
     return (
         <HeaderStyled>
             <Logo>$tock<span>market</span></Logo>
-            <SearchResult/>
+            <SearchForm/>
             <Navigation/>
         </HeaderStyled>
     )
