+++ src/components/SearchResult.jsx
@@ -1,28 +1,19 @@
-import React, {useRef, useCallback} from 'react';
-import {useHistory} from 'react-router-dom';
-
-
+import React from 'react';
 
 export default function SearchResult() {
-  const inputSearch = useRef();
-  const history = useHistory();
+  
 
-  const handleOnClick = useCallback(() => {
-    console.log(inputSearch);//TEST
-    console.log(inputSearch.current.value);//TEST
-    history.push('/search?q=' + encodeURIComponent(inputSearch.current.value));
-  }, [history]);
+  /* 
+  if (XXXXXX.includes(inputSearch.current.value)) {
+   console.log('This is a result') 
+  } else {
+    console.log('Item did not find')
+  } 
+  */
 
-
-    return (
-      <>
-        <input
-          ref = {inputSearch}
-          type='search'
-          placeholder='Search'
-        />
-        <button onClick={handleOnClick}>Search</button>
-         {/* <Link to="/search"></Link> */}
-      </>
-    );
-  }
+  return (
+    <>
+      
+    </>
+  );
+}
